fix: _builder_attribute_setter
